Squashed commit of the following:
commit 403343824bc2c330e0d96385e98c2da0640830b7
Author: pStyl3 <qp-game@gmx.de>
Date:   Thu May 9 22:46:53 2024 +0200

    Update German translation (#1970)

commit bc5a440a3d58425e27787f2a34f0fc6dab762d85
Author: Oliver Hamlet <git@ortham.net>
Date:   Wed May 8 17:49:55 2024 +0100

    Add timeouts to network requests

    The default value is 30 seconds.

commit 67c7fd5feb15148f67a2be55cfc466f1773acaa2
Author: Oliver Hamlet <git@ortham.net>
Date:   Wed May 8 17:49:18 2024 +0100

    Fix handling total failures of masterlist updates

    Since they involve multiple tasks, the handler is executed when the thread's execution ends, which could be due to the first task erroring.

commit b70691a527bfb9c3a3b8bdf22f6741c2bea23583
Author: Oliver Hamlet <git@ortham.net>
Date:   Tue May 7 17:24:08 2024 +0100

    Add errors for unsupported light plugins

commit 4aaceaeb0ce745d47545f741011a661251983221
Author: Oliver Hamlet <git@ortham.net>
Date:   Mon May 6 18:19:59 2024 +0100

    Fix accuracy of light plugin support check

    Skyrim VR and Fallout 4 VR do not support light plugins. However, there is an SKSEVR plugin that adds support for them, so check for it.

commit d63d9ca469f0cdd1f3957af959178848cb91e6a3
Author: Oliver Hamlet <git@ortham.net>
Date:   Sun May 5 19:23:00 2024 +0100

    Add plugin error if it is its own master

commit b8b4861a040f4f26c8bd486407b2437f8d7cfcf3
Author: Oliver Hamlet <git@ortham.net>
Date:   Sun May 5 18:58:04 2024 +0100

    Add config option to control case-sensitive paths warning

    It defaults to enabled and if disabled will stop LOOT from warning when running for a game that has case-sensitive local and/or install paths.

commit 367fbde4d77e82b6a6e6d4a470b64bb6bf21c108
Author: Oliver Hamlet <git@ortham.net>
Date:   Sun May 5 18:30:35 2024 +0100

    Add plugin card icons to docs

commit 5cfdc23a74cdbaec9283e95669e551cd5a17602e
Author: Oliver Hamlet <git@ortham.net>
Date:   Sun May 5 15:52:40 2024 +0100

    Update instructions for depending on a libloot snapshot

commit 4a83bc7dafc16b154b3b1660c90c3621a7fadc42
Author: Oliver Hamlet <git@ortham.net>
Date:   Sun May 5 15:43:34 2024 +0100

    Fix groups graph background colour

    It was previously documented as having no brush, but Qt 6.7 seems to have actually implemented that.

commit e908a5368c18e075ad076e45ffe1648ffa7c1e76
Author: Oliver Hamlet <git@ortham.net>
Date:   Fri May 3 21:15:37 2024 +0100

    Fix cached libloot missing from LD_LIBRARY_PATH in CI

commit 8c3c7e4f55ec98827a96920685bc2cbb85b89f19
Author: Oliver Hamlet <git@ortham.net>
Date:   Wed May 1 18:59:32 2024 +0100

    Fix building Flatpak in Ubuntu 22.04 in WSL2

    - Use LF line ending for Linux scripts even on Windows
    - Better handling of paths that include spaces
    - Avoid new f-string syntax to support Python 3.10
    - Replace shell substitution syntax that dash doesn't support

commit f4cc876c943e454f7da370f558892240914db20d
Author: Oliver Hamlet <git@ortham.net>
Date:   Sun Apr 28 16:43:48 2024 +0100

    Don't build Linux archives in CI

    Since the Flatpak can now be built from source, there's no longer any need for a release archive, and so there's no need to build the archive in CI builds either, so the Linux CI build just needs to run the tests.

commit e89c0024f54e61c7df59ebddd14ca52d60682d58
Author: Oliver Hamlet <git@ortham.net>
Date:   Fri Apr 26 21:58:12 2024 +0100

    Build Flatpak from source

    The Flatpak is built in a separate CI job because it takes about twice
    as long as the Linux job.

    There's a new script that generates manifests for the Python and Rust
    dependencies, as the Flatpak builder needs to download them ahead of
    time but can't just run pip or cargo itself. The script must be run
    before the Flatpak is built, and when the Flathub listing is updated the
    dependency manifests will need to be committed to its repository since
    Flathub's CI can't/won't run the script itself.

    This includes a few improvements to how the Flatpak is built:

    - --install-deps-from is used to avoid repeating the dependency versions
      unnecessarily
    - Built modules are cached, since some now take several minutes.
    - --ccache is used to enable caching of compiled object files, so even
      partial build results can be reused if their sources haven't changed.
    - The CI job now calls the helper script to reduce duplication.

commit 768d089cb5c23c6c42a86e2abb4a34569d314aea
Author: Oliver Hamlet <git@ortham.net>
Date:   Wed May 1 20:08:29 2024 +0100

    Allow expected hashes to be overridden

    They're tied to the URL as the hash shouldn't change if the URL hasn't, but if the URL has, then you might be trying to use a different version with a different hash.

commit 3e68fcf31702e2afcc326462bb346b8d1f7e3654
Author: Oliver Hamlet <git@ortham.net>
Date:   Fri May 3 19:12:04 2024 +0100

    Update Qt to 6.7.0

    This involves also updating ICU to 73.2 on Linux, as Qt 6.7.0 links
    against ICU 73. ICU 73 binaries are no longer available for Ubuntu
    20.04 and its glibc is too old to use with the 22.04 binaries, so
    update the runner to Ubuntu 22.04. That in turn required updating to GCC
    12 so that it could link with the updated TBB library installed by
    Ubuntu.

    This might cause runtime incompatibilities for the Flatpak, as I'm not
    sure what glibc the KDE 6.7 platform provides, but that possibility will
    be addressed in an upcoming commit.

commit 8ee3ed1a66b16cddef321a227a4493314398bab4
Author: Oliver Hamlet <git@ortham.net>
Date:   Fri May 3 17:28:58 2024 +0100

    Build libloot from source on Linux in GitHub Actions

    This will help avoid incompatibilities in ICU library versions between libloot, LOOT and Qt as they are updated.

commit 19efebaa021ef65cfbdd71736cbf7c3235d87c4a
Author: Oliver Hamlet <git@ortham.net>
Date:   Fri May 3 17:24:18 2024 +0100

    Update LD_LIBRARY_PATH when getting ICU

commit e6cc4c45e6f41bfe7f28659d33ad7d7d998b6ddd
Author: Oliver Hamlet <git@ortham.net>
Date:   Fri May 3 17:22:41 2024 +0100

    Update Boost to 1.85.0

commit 809121dc7581db1e9b05ca78c3839ecad4a5fd81
Author: Oliver Hamlet <git@ortham.net>
Date:   Fri May 3 17:21:00 2024 +0100

    Update ValveFileVDF to c8adfc29e62cc980b595e965bedfb239087647ff

commit b6ce22df8e3ee0069472451abb6926e36bf636d7
Author: Oliver Hamlet <git@ortham.net>
Date:   Fri May 3 17:20:31 2024 +0100

    Update spdlog to 1.14.1

commit 2d77e52d75f4961dd0439c1870632788f8d8ddba
Author: Oliver Hamlet <git@ortham.net>
Date:   Fri May 3 17:19:58 2024 +0100

    Update tomlplusplus to 3.4.0

commit 50180965365d93c519ecefb6b37396be35a54123
Author: Oliver Hamlet <git@ortham.net>
Date:   Fri May 3 17:19:30 2024 +0100

    Minizip fixup

commit 69eae99928793bf71d5bca4b11afde0cf06bebfa
Author: Oliver Hamlet <git@ortham.net>
Date:   Fri May 3 17:19:14 2024 +0100

    Update zlib to 1.3.1

commit 2e3fce2587221fe6721bd89c7a850c6752576de9
Author: Oliver Hamlet <git@ortham.net>
Date:   Fri May 3 17:18:57 2024 +0100

    Update minizip-ng to 4.0.5

commit c0e1706822293e3d82a9160b1e88cb6e40442d72
Author: Oliver Hamlet <git@ortham.net>
Date:   Fri May 3 17:18:14 2024 +0100

    Update libloot to 0.22.4

+++ src/gui/qt/tasks/network_task.cpp
@@ -51,7 +51,7 @@ void NetworkTask::onNetworkError(QNetworkReply::NetworkError networkError) {
     const auto logger = getLogger();
     if (logger) {
       logger->error("Network error code {}, description is: {}",
-                    networkError,
+                    static_cast<int>(networkError),
                     errorString);
     }
